requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
twilio==8.13.0
pandas==1.3.4


